+++ test/isolate_agents_test.dart
@@ -1,3 +1,5 @@
+import 'dart:isolate';
+
 import 'package:isolate_agents/isolate_agents.dart';
 import 'package:test/test.dart';
 
@@ -17,7 +19,25 @@ void main() {
 
   test('query', () async {
     final Agent<int> agent = await Agent.create(1);
-    final String value = await agent.query((state) => String.fromCharCode(65 + state));
+    final String value =
+        await agent.query((state) => String.fromCharCode(65 + state));
     expect('B', value);
+  });
+
+  test('hop agents', () async {
+    final Agent<int> agent1 = await Agent.create(1);
+    final Agent<int> agent2 = await Agent.create(2);
+    ReceivePort receivePort = ReceivePort();
+    SendPort sendPort = receivePort.sendPort;
+    agent1.send((x) {
+      agent2.send((y) {
+        sendPort.send(x + y);
+        return y;
+      });
+      return x;
+    });
+
+    int result = (await receivePort.first) as int;
+    expect(3, result);
   });
 }